~ Collider(1)

+++ GameBasic/PathManager.cpp
@@ -17,7 +17,7 @@ void PathManager::init()
 {
 	GetCurrentDirectory(255, contentPath);	
 
-	int len = wcslen(contentPath);
+	int len = (int)wcslen(contentPath);
 	for (int i = len - 1; 0 <= i; --i) {
 		if ('\\' == contentPath[i]) {
 			contentPath[i] = '\0';
